requirements.txt
botometer==1.6.1
geopy==2.2.0
numpy==1.21.5
pandas==1.4.2
PyYAML==6.0
requests==2.25.1
requests_oauthlib==1.3.1
tqdm==4.64.1
tweepy==3.10.0
